LRSD-4238 Regen
