+++ workspaces/liferay-testray-workspace/modules/testray-rest-api/src/main/java/com/liferay/testray/rest/resource/v1_0/TestrayTestFlowResource.java
@@ -12,7 +12,6 @@ import com.liferay.portal.odata.sort.SortParserProvider;
 import com.liferay.portal.vulcan.accept.language.AcceptLanguage;
 import com.liferay.portal.vulcan.batch.engine.resource.VulcanBatchEngineExportTaskResource;
 import com.liferay.portal.vulcan.batch.engine.resource.VulcanBatchEngineImportTaskResource;
-import com.liferay.testray.rest.dto.v1_0.TestrayCaseResultTestFlowUpdate;
 import com.liferay.testray.rest.dto.v1_0.TestrayTestFlow;
 
 import java.util.Collections;
@@ -42,10 +41,8 @@ import org.osgi.annotation.versioning.ProviderType;
 @ProviderType
 public interface TestrayTestFlowResource {
 
-	public TestrayCaseResultTestFlowUpdate
-			putTestrayTestFlowCaseResultByTestraySubtaskIdTestraySubtask(
-				Long testraySubtaskId,
-				TestrayCaseResultTestFlowUpdate testrayCaseResultTestFlowUpdate)
+	public TestrayTestFlow putTestrayTestFlowByTestraySubtaskIdTestraySubtask(
+			Long testraySubtaskId, TestrayTestFlow testrayTestFlow)
 		throws Exception;
 
 	public TestrayTestFlow postTestrayTestFlow(Long testrayTaskId)
